Streamline the Payload type a bit. Assorted cleanup and refactoring.
