+++ src/testrender/cuda/quad.cu
@@ -5,8 +5,8 @@
 
 #include <optix.h>
 
+#include "optix_raytracer.h"
 #include "rend_lib.h"
-#include "render_params.h"
 #include "vec_math.h"
 #include "wrapper.h"
 
@@ -51,7 +51,7 @@ __intersection__quad()
     // Check for self-intersection
     Payload payload;
     payload.get();
-    OSL_CUDA::ShaderGlobals* sg_ptr = (OSL_CUDA::ShaderGlobals*)payload.ptr.ptr;
+    OSL_CUDA::ShaderGlobals* sg_ptr = (OSL_CUDA::ShaderGlobals*)payload.sg_ptr;
     uint32_t* trace_data            = (uint32_t*)sg_ptr->tracedata;
     const int hit_idx               = ((int*)trace_data)[2];
     const int hit_kind              = ((int*)trace_data)[3];
